+++ src/shape.cpp
@@ -26,6 +26,8 @@
 #include <migraphx/sym.hpp>
 #include <migraphx/stringutils.hpp>
 #include <migraphx/serialize.hpp>
+#include <migraphx/json.hpp>
+#include <migraphx/convert_to_json.hpp>
 #include <migraphx/permutation.hpp>
 #include <migraphx/ranges.hpp>
 #include <numeric>
@@ -1493,6 +1495,11 @@ void migraphx_from_value(const value& v, shape& s)
             }
         }
     }
+}
+
+shape make_json_shape(const std::string& s)
+{
+    return from_value<shape>(from_json_string(convert_to_json(s)));
 }
 
 } // namespace MIGRAPHX_INLINE_NS